move PropertyValue types to shared
